Add preview label for current font selection
The currently selected font is previewed so that it does not
have to be added to the list of previews before the sample
is actually shown to the user. This involved moving some
parts of the interface, and adding two labels: one for the
currently selected font name, and one for the sample text in
the requested font and colors.

The glade files where editing using Glade 3, so there are
quite some (textual) changes there.

svn path=/branches/import-from-bzr/; revision=189